Option to discard bootstrap injector
Using child injectors can be problematic in certain situations where
the injector must be injected (usually part of framework code with
dynamic class loading) resulting in the wrong injector being injected.
While there are some workarounds to this problem they are not obvious
and have issues of their own.  This pull request make it possible to
have the LifecycleInjector discard the bootstrap injector thereby
making the injector returned by createInjector() a root injector.