Adding concept of an application runner
Currently it is necessary to manually start and stop the
LifecycleManager from outside of Governator.  This pull requests adds
the concept of an application runner that will handle the runtime
lifecycle of an entire application (i.e starting LifecycleManager,
waiting for the application to exit, closing LifecycleManager) as part
of the LifecycleInjector bootstrapping and allow for much simpler
integration with the container in which Governator is used (i.e
Standalone, Jetty, Tomcat, …).    A specific runner implementation is
implemented as a BootstrapModule which must provide a binding to
LifecycleRunner.

I also prototyped a simple Jetty based runner and would love to add it
as a sub project but governor has not been set up for sub projects yet.
 Any suggesting for how to proceed would be greatly appreciated.