Fix copying of bindings in simulated child injector mode
The previous code did not copy all bindings over to the simulated child
injector.  This new code iterates through all bindings, explicitly and
JIT, and copy all bindings except for a few core Guice specific
bindings that must be created in the new injector.  We may want to add
an exclude mechanism to this.