Add concept of a LifecycleInjectorBuilderSuite
The LifecycleInjectorBuilderSuite encapsulates a set of operations on
the LifecycleInjectorBuilder.  For now, suites are primarily used for
unit testing with the new LifecycleTester.  The LifecycleTester
simplifies testing since it can be used as a JUnit @Rule and can be
configured using suites and will auto shutdown the LifecycleManager at
the end of the test.