Fix bug with static resources on embedded Jetty
If there are multiple non-jar directories in the classpath then
whichever is scanned first by the classloader will show up as a
base resource. This complicates static resource management especially
when we want to do smoke tests or integration tests because we
need to have copies of the static resources (that we have under
src/main). That's not ideal.

Given these limitations in Jetty (and to avoid having to write custom
code to point to the app-specific static files (if we throw them under
META-INF/resources), we'll revert to the convention of having the app
specific static files under main/webapp.