Avoid referring to internal Guice classes
In Guice 4.2, guice-multibindings was merged into guice core, and in Guice 5, multibindings was deleted. When trying to upgrade to Guice5, I tried explicitly removing the multibindings dep as it is not rolled in, but some internal classes are no longer exposed.

In order to ease upgrading, Do a runtime check of the CircularDependency class, which was removed in Guice 5 ( https://github.com/google/guice/pull/1298/files ).